server/requirements.txt
pyserial
bcrypt
